requirements.txt
Django==4.2.16
pymysql
djangorestframework
django-filter
openpyxl
Crypto
pycryptodome
psycopg2
dotenv
kafka-python